moved shared compression functionality to compression_utils
